# Simplify common/locks.Lock, 5-10x speedup