Yeah, that code was just as bonkers as it looked.
    
Benchmark before:
```
BenchmarkLock-8                          2719692               450.3 ns/op           148 B/op          3 allocs/op
BenchmarkStdlibLock-8                   89316176                13.40 ns/op            0 B/op          0 allocs/op
BenchmarkParallelLock-8                  2119658               562.3 ns/op           148 B/op          3 allocs/op
BenchmarkParallelStdlibLock-8           10159339               111.6 ns/op             0 B/op          0 allocs/op
```
vs after:
```
BenchmarkLock-8                         37298678                32.16 ns/op            0 B/op          0 allocs/op
BenchmarkStdlibLock-8                   88782296                13.41 ns/op            0 B/op          0 allocs/op
BenchmarkParallelLock-8                  7922905               150.4 ns/op             0 B/op          0 allocs/op
BenchmarkParallelStdlibLock-8            9955582               113.0 ns/op             0 B/op          0 allocs/op
```
which is a _much_ more reasonable cost to pay for mutexes with timeouts.

I would entirely believe it's possible to improve further, but at the moment I'll (strongly) favor simplicity and obvious-correctness over peak performance, particularly since we apparently didn't need peak performance before.